Moved ScannerStatus into model folder, refactoring old implementation. Removed logging in ScannerController. Refactored IKDW and implementation, removing the ScannerController as constructor param for now: we have no specific configuration for it for now.
