+++ kotlin-data-wedge-lib/src/main/java/dk/gls/kdw/IKDW.kt
@@ -7,7 +7,7 @@ import dk.gls.kdw.model.barcode.ScannerConfiguration
 
 interface IKDW {
 
-    fun configure(context: Context, profileConfiguration: ProfileConfiguration, scannerController: ScannerController)
+    fun configure(context: Context, profileConfiguration: ProfileConfiguration)
 
     val scannerController: ScannerController
