--- kotlin-data-wedge-lib/src/main/java/dk/gls/kdw/configuration/scanner/ScannerStatus.kt
@@ -1,14 +1,0 @@
-package dk.gls.kdw.configuration.scanner
-
-enum class ScannerStatus(val value: String) {
-    WAITING("WAITING"),
-    SCANNING("SCANNING"),
-    CONNECTED("CONNECTED"),
-    DISCONNECTED("DISCONNECTED"),
-    IDLE("IDLE"),
-    DISABLED("DISABLED")
-}
-
-fun String.toScannerStatusEnum(): ScannerStatus {
-    return ScannerStatus.valueOf(this)
-}
