+++ kotlin-data-wedge-lib/src/main/java/dk/gls/kdw/model/scanner/ScannerStatus.kt
@@ -1,6 +1,14 @@
 package dk.gls.kdw.model.scanner
 
-sealed interface ScannerStatus {
-    object Enabled : ScannerStatus
-    object Disabled : ScannerStatus
+enum class ScannerStatus(val value: String) {
+    WAITING("WAITING"),
+    SCANNING("SCANNING"),
+    CONNECTED("CONNECTED"),
+    DISCONNECTED("DISCONNECTED"),
+    IDLE("IDLE"),
+    DISABLED("DISABLED")
+}
+
+fun String.toScannerStatusEnum(): ScannerStatus {
+    return ScannerStatus.valueOf(this)
 }
